+++ test/coinstring.test.js
@@ -9,9 +9,15 @@ describe('coinstring', function() {
       it('should encode ' + f.description, function() {
         var inp = new Buffer(f.hex, 'hex')
         var version = parseInt(f.version, '16')
+
+        if (f.version.length < 4)  {//skip bip32 here
+          assert.equal(cs.encode(inp, version), f.base58)
+          assert.equal(cs.encode([].slice.call(inp), version), f.base58)
+          assert.equal(cs.encode(new Uint8Array([].slice.call(inp)), version), f.base58)
+        }
+
+        version = new Buffer(f.version.substr(2), 'hex') //chop off '0x'
         assert.equal(cs.encode(inp, version), f.base58)
-        assert.equal(cs.encode([].slice.call(inp), version), f.base58)
-        assert.equal(cs.encode(new Uint8Array([].slice.call(inp)), version), f.base58)
       })
     })
   })
@@ -20,18 +26,31 @@ describe('coinstring', function() {
     fixtures.valid.forEach(function(f) {
       it('should decode ' + f.description, function() {
         var version = parseInt(f.version, '16')
-        assert.equal(cs.decode(f.base58, version).payload.toString('hex'), f.hex)
-        assert.equal(cs.decode(f.base58).payload.toString('hex'), f.hex)
+        var bufferVersion = new Buffer(f.version.substr(2), 'hex')
+        var res = {}
+
+        if (f.version.length < 4) { //skip bip32
+          res = cs.decode(f.base58, version)
+          assert.equal(res.payload.toString('hex'), f.hex)
+          assert.equal(res.version.toString('hex'), f.version.substr(2))
+          res = cs.decode(f.base58)
+          assert.equal(res.payload.toString('hex'), f.hex)
+          assert.equal(res.version.toString('hex'), f.version.substr(2))
+        } 
+
+        res = cs.decode(f.base58, bufferVersion)
+        assert.equal(res.payload.toString('hex'), f.hex)
+        assert.equal(res.version.toString('hex'), f.version.substr(2))
       })
     })
 
     describe('> when invalid input', function() {
       fixtures.invalid.forEach(function(f) {
-        it(f.description + ' should throw an error', function() {
+        it(f.description + ': ' + f.base58 + ' should throw an error', function() {
           assert.throws(function() {
             var version = parseInt(f.version, '16')
-            assert(cs.decode(version, f.base58))
-          })
+            assert(cs.decode(f.base58, version))
+          },new RegExp(f.match, 'i'))
         })
       })
     })
@@ -41,7 +60,10 @@ describe('coinstring', function() {
     fixtures.valid.forEach(function(f) {
       it('should validate ' + f.description, function() {
         var version = parseInt(f.version, '16')
-        assert(cs.isValid(f.base58, version))
+        var versionBuffer = new Buffer(f.version.substr(2), 'hex')
+        if (f.version.length < 4) //skip bip 32
+          assert(cs.isValid(f.base58, version))
+        assert(cs.isValid(f.base58, versionBuffer))
       })
     })
 
@@ -60,8 +82,14 @@ describe('coinstring', function() {
       it('should create an encoder ' + f.description, function() {
         var inp = new Buffer(f.hex, 'hex')
         var version = parseInt(f.version, '16')
+        var versionBuffer = new Buffer(f.version.substr(2), 'hex')
 
-        var encode = cs.createEncoder(version)
+        if (f.version.length < 4) {//skip bip 32
+          var encode = cs.createEncoder(version)
+          assert.equal(encode(inp), f.base58)
+        }
+
+        encode = cs.createEncoder(versionBuffer)
         assert.equal(encode(inp), f.base58)
       })
     })
@@ -72,8 +100,14 @@ describe('coinstring', function() {
       it('should create a decoder ' + f.description, function() {
         var inp = new Buffer(f.hex, 'hex')
         var version = parseInt(f.version, '16')
+        var versionBuffer = new Buffer(f.version.substr(2), 'hex')
 
-        var decode = cs.createDecoder(version)
+        if (f.version.length < 4) { //skip bip 32
+          var decode = cs.createDecoder(version)
+          assert.equal(decode(f.base58).payload.toString('hex'), f.hex)
+        }
+
+        decode = cs.createDecoder(versionBuffer)
         assert.equal(decode(f.base58).payload.toString('hex'), f.hex)
       })
     })
@@ -83,7 +117,13 @@ describe('coinstring', function() {
     fixtures.valid.forEach(function(f) {
       it('should create a validator ' + f.description, function() {
         var version = parseInt(f.version, '16')
-        var isValid = cs.createValidator(version)
+        var versionBuffer = new Buffer(f.version.substr(2), 'hex')
+        if (f.version.length < 4) { //skip bip 32
+          var isValid = cs.createValidator(version)
+          assert(isValid(f.base58))
+        }
+
+        isValid = cs.createValidator(versionBuffer)
         assert(isValid(f.base58))
       })
     })
@@ -92,7 +132,14 @@ describe('coinstring', function() {
       fixtures.invalid.forEach(function(f) {
         it(f.description + ' should return false', function() {
           var version = parseInt(f.version, '16')
-          var isValid = cs.createValidator(version)
+          var versionBuffer = new Buffer(f.version.substr(2), 'hex')
+
+          if (f.version.length < 4) { //skip bip 32
+            var isValid = cs.createValidator(version)
+            assert(!isValid(f.base58))
+          }
+
+          isValid = cs.createValidator(versionBuffer)
           assert(!isValid(f.base58))
         })
       })