+++ CHANGELOG.md
@@ -16,6 +16,7 @@ x.y.z / 2014-06-dd
 * changed method signature of `validate(version, base58str)` to `validate(base58str, version)`
 * renamed method `validate()` to `isValid()`
 * renamed return from `decode()` object field `bytes` to `payload`
+* added `Buffer` support for version input => BIP32 support
 
 0.2.0 / 2014-03-10
 ------------------